Final version of encode/decode for PR review; separated program into individual files
